Split admin dashboard data helpers
